Move class LabelImageUtils and LabelImageUtilsTest to package org.mastodon.mamut.util
