+++ src/main/java/org/mastodon/mamut/io/importer/labelimage/ui/ImportSpotsFromBdvChannelView.java
@@ -30,7 +30,7 @@ package org.mastodon.mamut.io.importer.labelimage.ui;
 
 import bdv.viewer.SourceAndConverter;
 import org.mastodon.mamut.ProjectModel;
-import org.mastodon.mamut.io.importer.labelimage.LabelImageUtils;
+import org.mastodon.mamut.util.LabelImageUtils;
 import org.scijava.ItemVisibility;
 import org.scijava.command.Command;
 import org.scijava.command.DynamicCommand;
